+++ app/src/main/res/layout/activity_cadastro.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimaryDark"
-    tools:context=".Cadastro">
+    tools:context=".view.CadastroActivity">
 
     <ImageView
         android:id="@+id/logoTrunflix"
