+++ app/src/main/res/layout/activity_filtro.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/preto"
-    tools:context=".FiltroActivity">
+    tools:context=".view.FiltroActivity">
 
     <CheckBox
         android:id="@+id/checkFantasia"
