+++ app/src/main/res/layout/activity_selecao_jogo.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/preto"
     android:orientation="vertical"
-    tools:context=".SelecaoJogo">
+    tools:context=".view.SelecaoJogo">
 
 
     <LinearLayout
@@ -15,6 +15,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/vermelho"
+        android:layout_marginTop="16dp"
         android:orientation="horizontal">
 
         <TextView
