+++ app/src/main/res/layout/activity_trunfo.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/preto"
-    tools:context=".Trunfo">
+    tools:context=".view.TrunfoActivity">
 
     <TextView
         android:id="@+id/pontosPerdeu"
@@ -17,6 +17,7 @@
         android:text="1"
         android:textColor="@color/branco"
         android:textSize="40sp"
+        android:paddingBottom="2dp"
         app:layout_constraintBottom_toTopOf="@+id/recyclerCartas"
         app:layout_constraintEnd_toStartOf="@+id/setaPerdeu" />
 
@@ -95,6 +96,7 @@
         android:layout_marginBottom="8dp"
         android:background="@drawable/back_ganhou_perdeu"
         android:gravity="center"
+        android:paddingBottom="2dp"
         android:text="2"
         android:textColor="@color/branco"
         android:textSize="40sp"
