+++ app/src/main/res/layout/carta.xml
@@ -2,6 +2,7 @@
 <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/layoutCartaTrunfo"
     android:layout_width="340dp"
     android:layout_height="560dp"
     android:background="@color/vermelhoCarta"
