+++ app/src/main/res/layout/carta_quiz.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:background="@color/vermelhoCarta">
+
+
+    <android.support.v7.widget.CardView
+        android:layout_width="261dp"
+        android:layout_height="256dp"
+        android:background="@color/vermelhoCarta"
+
+        >
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="vertical">
+
+            <ImageView
+                android:id="@+id/imagemCartaQuiz"
+                android:layout_width="290dp"
+                android:layout_height="225dp"
+                android:layout_gravity="center_horizontal"
+                android:background="@color/vermelhoCarta"
+                android:src="@drawable/taxidriver"/>
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="36dp"
+                android:background="@color/preto"
+                android:orientation="horizontal">
+
+                <TextView
+                    android:id="@+id/tituloCartaQuiz"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:textColor="@color/vermelho"
+                    android:textSize="15sp"
+                    tools:text="Taxi Driver" />
+
+                <TextView
+                    android:id="@+id/anoCartaQuiz"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginLeft="6dp"
+                    android:paddingTop="5dp"
+                    android:textColor="@color/vermelho"
+                    android:textSize="15dp"
+                    tools:text="2000" />
+            </LinearLayout>
+
+
+
+        </LinearLayout>
+
+    </android.support.v7.widget.CardView>
+
+
+</LinearLayout>
