+++ app/src/main/res/values/colors.xml
@@ -8,5 +8,5 @@
     <color name="vermelho">#DB0000</color>
     <color name="vermelhoCarta">#CE1316</color>
     <color name="branco">#FFF</color>
-    <color name="cinza_input">#5A6978</color>
+    <color name="cinza_input">#888A92</color>
 </resources>
