+++ app/src/main/res/values/strings.xml
@@ -44,4 +44,5 @@
     <string name="escolha_as_d_cadas">ESCOLHA AS DÉCADAS:</string>
     <string name="_70_s">70\'s</string>
     <string name="_90_s">90\'s</string>
+    <string name="qual_dos_filmes_teve_a_maior_bilheteria">QUAL DOS FILMES TEVE A MAIOR BILHETERIA ?</string>
 </resources>
